+++ R/restore.R
@@ -406,40 +406,11 @@ installPkg <- function(pkgRecord,
     }
   }
 
-  # Generally we want to install from sources, but we will download a pre-
-  # built binary if (a) the package exists on CRAN, (b) the version on CRAN
-  # is the version desired, and (c) R is set to download binaries. We also
-  # just copy a symlink from the cache if possible.
-  copiedFromCache <- FALSE
-  if (isUsingCache(project) &&
-      length(pkgRecord$hash) &&
-      isCacheable(pkgRecord$name) &&
-      pkgRecord$hash %in% cache) {
-
-    # This package has already been installed in the cache; just symlink
-    # to that if possible, or copy otherwise
-    success <- suppressWarnings(symlink(
-      file.path(cacheLibDir(pkgRecord$name, pkgRecord$hash, pkgRecord$name)),
-      file.path(libDir(project), pkgRecord$name)
-    ))
-
-    if (success) {
-      type <- "symlinked cache"
-      needsInstall <- FALSE
-    } else {
-      ## just copy the directory over instead of symlinking
-      success <- all(dir_copy(
-        file.path(cacheLibDir(pkgRecord$name, pkgRecord$hash)),
-        file.path(libDir(project), pkgRecord$name)
-      ))
-      if (!success) {
-        warning("Failed to symlink or copy package '", pkgRecord$name, "' from cache")
-      } else {
-        type <- "copied cache"
-        needsInstall <- FALSE
-      }
-    }
-    copiedFromCache <- success
+  cacheCopyStatus <- new.env(parent = emptyenv())
+  copiedFromCache <- restoreWithCopyFromCache(project, pkgRecord, cache, cacheCopyStatus)
+  if (copiedFromCache) {
+    type <- cacheCopyStatus$type
+    needsInstall <- FALSE
   }
 
   if (!(copiedFromCache) &&